MetrocardBonusCalculator
This programs takes the amount of money left on a user's New York City
Metrocard and the maximum amount of money that they are willing to
spend. Then it calculates all possibilities of amounts that can be
purchased so the user ends up with an even number of rides after the 5%
bonus is applied.